createDraftOrder& sendDraftOrder

+++ model/orderItem.go
@@ -7,7 +7,7 @@ type OrderItem struct {
 
 	ProductName string   `gorm:"type:varchar(255);not null" json:"productName"`
 	SizeInch    *float64 ` json:"sizeInch"`                     // kích thước (inch)
-	Quantity    *float64 ` json:"quantity"`                     // số lượng
+	Quantity    *int     ` json:"quantity"`                     // số lượng
 	Unit        *string  `gorm:"type:varchar(20);" json:"unit"` // đơn vị: wig, bundles, pieces, kg
 	UnitPrice   *float64 ` json:"unitPrice"`                    // USD
 	Subtotal    *float64 ` json:"subtotal"`                     // = Quantity * UnitPrice
